requirements.txt
accelerate==1.10.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.89.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
boto3==1.42.83
botocore==1.42.83
bracex==2.6
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
dataclasses-json==0.6.7
deepagents==0.4.12
Deprecated==1.2.18
distro==1.9.0
docstring_parser==0.17.0
effdet==0.4.1
emoji==2.14.1
eval_type_backport==0.2.0
faiss-cpu==1.11.0.post1
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-auth==2.49.1
google-cloud-vision==3.10.2
google-genai==1.63.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
hf-xet==1.1.7
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
iniconfig==2.3.0
iopath==0.1.10
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==1.2.15
langchain-anthropic==1.4.0
langchain-aws==1.4.3
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.26
langchain-google-genai==4.2.0
langchain-ollama==0.3.10
langchain-text-splitters==1.1.1
langdetect==1.0.9
langgraph==1.1.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.4.13
layoutparser==0.3.4
lxml==6.0.0
Markdown==3.8.2
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.2.6
olefile==0.47
ollama==0.6.1
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.1
opencv-python==4.12.0.88
orjson==3.11.8
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pi_heif==1.1.0
pikepdf==9.10.2
pillow==11.3.0
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
portalocker==3.0.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.3
pypdf==5.9.0
pypdfium2==4.30.0
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
redis==4.6.0
referencing==0.37.0
regex==2025.7.34
requests==2.33.1
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
safetensors==0.6.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
sqlite-vec==0.1.6
sympy==1.14.0
tenacity==9.1.2
tesseract==0.1.3
tiktoken==0.8.0
timm==1.0.19
tokenizers
torch
torchvision
tqdm==4.67.1
transformers==4.55.0
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unstructured==0.18.11
unstructured-client==0.42.2
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3==2.5.0
uuid_utils==0.14.0
wcmatch==10.1
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
xxhash==3.6.0
yarl==1.20.1
zstandard==0.23.0
